better user-defined background picture
git-svn-id: http://subversion.assembla.com/svn/WebAlbums@231 373d8ef0-abc8-4151-9680-7fa2cb6b1372

+++ WebAlbums3/trunk/WebAlbums3-Service-Beans/src/java/net/wazari/service/engine/PhotoBean.java
@@ -151,9 +151,11 @@ public class PhotoBean implements PhotoLocal {
             File backgroundDir = new File(vSession.getConfiguration()
                     .getTempPath()+vSession.getTheme().getNom()) ;
             log.info("Delete and create background dir: {}", backgroundDir) ;
-            for (File child : backgroundDir.listFiles()) {
-                if (!child.delete()) {
-                    log.warn("Could not remove background file: {}",child);
+            if (backgroundDir.listFiles() != null) {
+                for (File child : backgroundDir.listFiles()) {
+                    if (!child.delete()) {
+                        log.warn("Could not remove background file: {}",child);
+                    }
                 }
             }
         }
